+++ src/main/java/lotto/exception/LottoException.java
@@ -1,5 +1,7 @@
 package lotto.exception;
 
+import lotto.exception.constants.ErrorMessage;
+
 public class LottoException extends IllegalArgumentException {
     private LottoException(final ErrorMessage errorMessage) {
         super(errorMessage.getMessage());
